requirements.txt
flask
flask-sqlalchemy
groq
python-dotenv
chromadb
sentence-transformers
pypdf2
markdown
werkzeug